And we have tests. Also, move error logging upstream.
This is both to make the tests more robust and also because I'll want to
surface these kind of logs to the client once I get that going. I don't
want to have to go fishing through the server's /var/log output for
issues with an RSS feed. That'll be only for really serious errors with
the application itself, not things I expect will fail.